Move tour.py to app namespace
