+++ spyderlib/app/tour.py
